+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 - **Forum Changes**
   - The forums have been simplified into a single forum
   - Users can jump to the last post on a topic from the forum index
+  - Forum post and reply interface has been updated
 - **General Changes**
   - Simplified user interface and styling across site
 - **Menu Changes**